Add files via upload
Analytics Project where I read in a CSV file and perform Simple Linear Regression and Residual Analysis on two variables.  I interpret the coefficients found and what they mean for the variables in my analysis. I created a scatterplot comparing the variables which includes a exponential line of best fit with the standard error. Also created a scatterplot of the residual values and their distance from the least squares line. I then log-transformed the variables and made interpretations on the coefficients in a simple linear regression analysis and a residual analysis.