Add files via upload
The data set HMEQ reports characteristics and delinquency information for 5,960 home equity loans. A home equity loan is a loan where the obligor uses the equity of his or her home as the underlying collateral. I built five different analytical models and chose the best one to use based on the results found.